Add files via upload
Project Overview:
Welcome to the Titanic Survival Prediction project! This repository contains the code and resources for predicting the survival status of passengers aboard the infamous RMS Titanic using machine learning techniques. The sinking of the Titanic is one of the most well-known maritime disasters in history, and this project aims to analyze the available data to determine the factors that influenced survival rates.

Project Goals:
The primary goals of this project are as follows:

Data Exploration: Perform an in-depth exploration of the Titanic dataset to gain insights into the characteristics of passengers and understand the distribution of survivors.
Feature Engineering: Preprocess the data and engineer relevant features to improve the performance of the machine learning models.
Machine Learning Models: Implement various classification algorithms to predict the survival status of passengers accurately.
Model Evaluation: Evaluate the performance of different models using appropriate metrics and compare their results.
Model Deployment: (Optional) Deploy the best-performing model as a web service to enable real-time predictions.
Project Structure:
The repository is organized into the following directories:

data: Contains the Titanic dataset in CSV format.
notebooks: Jupyter notebooks used for data exploration, feature engineering, and model development.
src: Python scripts for utility functions, data preprocessing, and model training.
models: Saved model files or checkpoints if applicable.
images: Image files for documentation and README.md.
requirements.txt: Lists the required dependencies for easy setup.
LICENSE: License information for the project.
Data Source:
The dataset used in this project is sourced from the well-known Kaggle competition "Titanic: Machine Learning from Disaster." It consists of passenger information such as age, gender, ticket class, cabin, fare, and whether they survived or not.